cleanup smt benches
