+++ base_layer/mmr/benches/smt.rs
@@ -19,7 +19,7 @@ fn create_smt() -> SparseMerkleTree<Blake2b<U32>> {
     SparseMerkleTree::<Blake2b<U32>>::new()
 }
 
-pub fn benchmark_sparse_merkle_trees(c: &mut Criterion) {
+pub fn benchmark_smt_insert(c: &mut Criterion) {
     let sizes = [100, 10_000];
     for size in sizes {
         c.bench_function(&format!("SMT: Insert {size} keys"), move |b| {
@@ -38,5 +38,57 @@ pub fn benchmark_sparse_merkle_trees(c: &mut Criterion) {
     }
 }
 
-criterion_group!(smt, benchmark_sparse_merkle_trees);
+fn insert_into_smt(keys: &[NodeKey], tree: &mut SparseMerkleTree<Blake2b<U32>>) {
+    keys.iter().for_each(|key| {
+        tree.upsert(key.clone(), ValueHash::default()).unwrap();
+    });
+}
+
+fn delete_from_smt(keys: &[NodeKey], tree: &mut SparseMerkleTree<Blake2b<U32>>) {
+    keys.iter().for_each(|key| {
+        tree.delete(key).unwrap();
+    });
+}
+
+fn time_function(header: &str, f: impl FnOnce()) -> std::time::Duration {
+    println!("Starting: {header}");
+    let now = std::time::Instant::now();
+    f();
+    let t = now.elapsed();
+    println!("Finished: {header} - {t:?}");
+    t
+}
+
+pub fn root_hash(_c: &mut Criterion) {
+    let size = 1_000_000;
+    let half_size = size / 2;
+    let keys = get_keys(size);
+    let mut tree = create_smt();
+    time_function(&format!("SMT: Inserting {size} keys"), || {
+        insert_into_smt(&keys, &mut tree);
+    });
+    time_function("SMT: Calculating root hash", || {
+        let size = tree.size();
+        let hash = tree.hash();
+        println!("Tree size: {size}. Root hash: {hash:x}");
+    });
+    time_function(&format!("SMT: Deleting {half_size} keys"), || {
+        delete_from_smt(&keys[0..half_size], &mut tree);
+    });
+    time_function("SMT: Calculating root hash", || {
+        let size = tree.size();
+        let hash = tree.hash();
+        println!("Tree size: {size}. Root hash: {hash:x}");
+    });
+    time_function(&format!("SMT: Deleting another {half_size} keys"), || {
+        delete_from_smt(&keys[half_size..], &mut tree);
+    });
+    time_function("SMT: Calculating root hash", || {
+        let size = tree.size();
+        let hash = tree.hash();
+        println!("Tree size: {size}. Root hash: {hash:x}");
+    });
+}
+
+criterion_group!(smt, benchmark_smt_insert, root_hash);
 criterion_main!(smt);
